notebooks/3D-point-pillars/requirements.txt
numba==0.62.1
numpy==2.2.6
open3d==0.19.0
opencv-python-headless==4.12.0.88
PyYAML==6.0.3
setuptools==80.9.0
tensorboard==2.20.0
torch==2.9.0
openvino==2025.3.0
tqdm==4.67.1
onnx==1.19.1
